🔨 chore: add check to throw NotAllowedError if the user is not allowed to delete the meal

+++ src/use-cases/meals/delete-meal-usecase.ts
@@ -1,19 +1,25 @@
 import { MealsRepository } from '@/repositories/meals/meals-repository'
 
 import { MealNotFoundError } from './errors/meal-not-found-error'
+import { NotAllowedError } from './errors/not-allowed-error'
 
 interface DeleteMealUseCaseRequest {
+  userId: string
   mealId: string
 }
 
 export class DeleteMealUseCase {
   constructor(private readonly mealsRepository: MealsRepository) {}
 
-  async execute({ mealId }: DeleteMealUseCaseRequest): Promise<void> {
+  async execute({ userId, mealId }: DeleteMealUseCaseRequest): Promise<void> {
     const meal = await this.mealsRepository.findById(mealId)
 
     if (!meal) {
       throw new MealNotFoundError()
+    }
+
+    if (meal.user_id !== userId) {
+      throw new NotAllowedError()
     }
 
     await this.mealsRepository.delete(meal.id)
